updated monte-python to correct version
